Change to Long Computer

+++ src/ch/feomathar/adventofcode/dayfive/SunnyAsteroids.kt
@@ -5,7 +5,7 @@ import ch.feomathar.adventofcode.util.parseInput
 
 fun diagnose(fileName: String): String {
     val code = parseInput(fileName)[0]
-    val computer = Computer(code.split(",").map { it.toInt() })
+    val computer = Computer(code.split(",").map { it.toLong() })
 //    computer.addInput("1")
     computer.addInput("5")
     computer.runProgram()
